refactor(workbooks): extract Format/Group/Columns panels from Grid.tsx (EP-GRID-WORKBOOKS)
Continues the WorkbookGrid decomposition begun in #3045 (Filter + Summary
panels). The three remaining inline toolbar panels — conditional-format
(Format), in-grid grouping (Group), and row-height/freeze/hide-fields
(Columns) — now live as typed, presentational components in GridPanels.tsx,
matching the GridFilterPanel/GridSummaryPanel pattern (pure UI over props the
Grid owns).

Pure refactor — the JSX is moved verbatim (the Format panel's local
`ctrlClass` was byte-identical to GridPanels' shared `CTRL`), so behavior,
markup, aria labels, and classNames are unchanged. No contract change.

Grid.tsx drops 1298 → 1094 LOC; GridPanels.tsx grows 298 → 610 (still under
the 800 size-guard threshold, so unbaselined). The panels were the growth
driver — every parity feature added one — so extracting them is what keeps
the Grid focused on data / react-data-grid wiring and lets the next panel
land as its own file. Removed the now-unused conditional-format and
ROW_HEIGHTS imports from Grid.tsx.

Local tsc skipped (source-only worktree: @dpf/db generated client absent);
CI runs the authoritative typecheck. Unit tests green (113/113 workbooks).

Design-Grounding-Decision: no-contract-change refactor of the apps/web/ grid substrate; continues the decomposition tracked in docs/superpowers/plans/2026-06-11-universal-grid-workbooks-phase2.md (Debt note). No spec/behavior change.
UX-Fit-Decision: no-op — identical UI, zero new or changed controls; the exact same JSX moved from Grid.tsx into GridPanels.tsx components.
Process-Spine-Decision: touched docs/superpowers/plans/2026-06-11-universal-grid-workbooks-phase2.md (Debt note); no new module (GridPanels.tsx already exists) and no new spec needed.
